requirements.txt
streamlit
ultralytics
torch
pillow
numpy
